feat: update CollectionCard component with `ConfirmationDialog`
